fix(vselect): fix for empty dense vSelect incorrect height
The dense input style on the empty text input of a dense mode VSelect is causing the element to be
too short. When an item is selected, the element visibly grows and returns to the proper size. By
exluding the dense padding override on the text field, the Empty Dense VSelect now matches other
dense form controls like autocomplete.

Fix #10635